Add settings to jekyll that will allow me to create the static blog for github pages
This isn't the most elegant workflow, but as it stands, create posts in tblog/_posts folder, then run the rails server. this will initialize with the blog.rb code, and create the static in public/blog.  In that folder, the beaukinstler github pages can be commited and pushed to the public site.
Perhaps this can be automoted in the future, but I think it will allow
me to play around a bit first.